+++ src/awkward/operations/ak_to_backend.py
@@ -1,6 +1,7 @@
 # BSD 3-Clause License; see https://github.com/scikit-hep/awkward-1.0/blob/main/LICENSE
 __all__ = ("to_backend",)
 import awkward as ak
+from awkward._backends.dispatch import regularize_backend
 from awkward._behavior import behavior_of
 from awkward._layout import wrap_layout
 from awkward._nplikes.numpylike import NumpyMetadata
@@ -66,5 +67,5 @@ def _impl(array, backend, highlevel, behavior):
         allow_other=True,
     )
     behavior = behavior_of(array, behavior=behavior)
-    backend_layout = layout.to_backend(ak._backends.regularize_backend(backend))
+    backend_layout = layout.to_backend(regularize_backend(backend))
     return wrap_layout(backend_layout, behavior, highlevel)
