+++ src/awkward/record.py
@@ -5,7 +5,8 @@ import copy
 from collections.abc import Iterable
 
 import awkward as ak
-from awkward._backends import Backend
+from awkward._backends.backend import Backend
+from awkward._backends.dispatch import regularize_backend
 from awkward._behavior import get_record_class
 from awkward._layout import wrap_layout
 from awkward._nplikes.numpylike import NumpyMetadata
@@ -228,7 +229,7 @@ class Record:
         if backend is None:
             backend = self._array.backend
         else:
-            backend = ak._backends.regularize_backend(backend)
+            backend = regularize_backend(backend)
         if backend is self._array.backend:
             return self
         else:
